Refactor introspection client into components for clarity
