+++ oak_runtime/introspection_browser_client/webpack.config.js
@@ -38,7 +38,10 @@ module.exports = env => ({
     ]
   },
   resolve: {
-    extensions: ['.ts', '.tsx', '.js']
+    extensions: ['.ts', '.tsx', '.js'],
+    alias: {
+      '~': path.resolve(__dirname)
+    }
   },
   mode: env.NODE_ENV || 'none',
   devServer: {
